backend/requirements.txt
django
web3
celery
django-celery-beat
django-celery-results
python-dotenv
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-cors-headers